split client.rs
