+++ h3/src/client/connection.rs
@@ -1,19 +1,17 @@
 //! Client implementation of the HTTP/3 protocol
 
 use std::{
-    convert::TryFrom,
     marker::PhantomData,
     sync::{atomic::AtomicUsize, Arc},
     task::{Context, Poll, Waker},
 };
 
-use bytes::{Buf, Bytes, BytesMut};
+use bytes::{Buf, BytesMut};
 use futures_util::future;
-use http::{request, HeaderMap, Response};
+use http::request;
 use tracing::{info, trace};
 
 use crate::{
-    config::Config,
     connection::{self, ConnectionInner, ConnectionState, SharedStateRef},
     error::{Code, Error, ErrorLevel},
     frame::FrameStream,
@@ -23,25 +21,7 @@ use crate::{
     stream::{self, BufRecvStream},
 };
 
-/// Start building a new HTTP/3 client
-pub fn builder() -> Builder {
-    Builder::new()
-}
-
-/// Create a new HTTP/3 client with default settings
-pub async fn new<C, O>(conn: C) -> Result<(Connection<C, Bytes>, SendRequest<O, Bytes>), Error>
-where
-    C: quic::Connection<Bytes, OpenStreams = O>,
-    O: quic::OpenStreams<Bytes>,
-{
-    //= https://www.rfc-editor.org/rfc/rfc9114#section-3.3
-    //= type=implication
-    //# Clients SHOULD NOT open more than one HTTP/3 connection to a given IP
-    //# address and UDP port, where the IP address and port might be derived
-    //# from a URI, a selected alternative service ([ALTSVC]), a configured
-    //# proxy, or name resolution of any of these.
-    Builder::new().build(conn).await
-}
+use super::stream::RequestStream;
 
 /// HTTP/3 request sender
 ///
@@ -62,7 +42,7 @@ where
 /// ## Sending a request with no body
 ///
 /// ```rust
-/// # use h3::{quic, client::client::*};
+/// # use h3::{quic, client::connection::*, client::stream::*};
 /// # use http::{Request, Response};
 /// # use bytes::Buf;
 /// # async fn doc<T,B>(mut send_request: SendRequest<T, B>) -> Result<(), Box<dyn std::error::Error>>
@@ -88,7 +68,7 @@ where
 /// ## Sending a request with a body and trailers
 ///
 /// ```rust
-/// # use h3::{quic, client::client::*};
+/// # use h3::{quic, client::connection::*};
 /// # use http::{Request, Response, HeaderMap};
 /// # use bytes::{Buf, Bytes};
 /// # async fn doc<T,B>(mut send_request: SendRequest<T, Bytes>) -> Result<(), Box<dyn std::error::Error>>
@@ -125,14 +105,14 @@ where
     T: quic::OpenStreams<B>,
     B: Buf,
 {
-    open: T,
-    conn_state: SharedStateRef,
-    max_field_section_size: u64, // maximum size for a header we receive
+    pub(super) open: T,
+    pub(super) conn_state: SharedStateRef,
+    pub(super) max_field_section_size: u64, // maximum size for a header we receive
     // counts instances of SendRequest to close the connection when the last is dropped.
-    sender_count: Arc<AtomicUsize>,
-    conn_waker: Option<Waker>,
-    _buf: PhantomData<fn(B)>,
-    send_grease_frame: bool,
+    pub(super) sender_count: Arc<AtomicUsize>,
+    pub(super) conn_waker: Option<Waker>,
+    pub(super) _buf: PhantomData<fn(B)>,
+    pub(super) send_grease_frame: bool,
 }
 
 impl<T, B> SendRequest<T, B>
@@ -283,7 +263,7 @@ where
 /// ```rust
 /// # use bytes::Buf;
 /// # use futures_util::future;
-/// # use h3::{client::client::*, quic};
+/// # use h3::{client::connection::*, quic};
 /// # use tokio::task::JoinHandle;
 /// # async fn doc<C, B>(mut connection: Connection<C, B>)
 /// #    -> JoinHandle<Result<(), Box<dyn std::error::Error + Send + Sync>>>
@@ -306,7 +286,9 @@ where
 /// ```rust
 /// # use bytes::Buf;
 /// # use futures_util::future;
-/// # use h3::{client::client::*, quic};
+/// # use h3::quic;
+/// # use h3::client::connection::Connection;
+/// # use h3::client::connection::SendRequest;
 /// # use tokio::{self, sync::oneshot, task::JoinHandle};
 /// # async fn doc<C, B>(mut connection: Connection<C, B>)
 /// #    -> Result<(), Box<dyn std::error::Error + Send + Sync>>
@@ -351,11 +333,11 @@ where
     C: quic::Connection<B>,
     B: Buf,
 {
-    inner: ConnectionInner<C, B>,
+    pub(super) inner: ConnectionInner<C, B>,
     // Has a GOAWAY frame been sent? If so, this PushId is the last we are willing to accept.
-    sent_closing: Option<PushId>,
+    pub(super) sent_closing: Option<PushId>,
     // Has a GOAWAY frame been received? If so, this is StreamId the last the remote will accept.
-    recv_closing: Option<StreamId>,
+    pub(super) recv_closing: Option<StreamId>,
 }
 
 impl<C, B> Connection<C, B>
@@ -460,302 +442,5 @@ where
         }
 
         Poll::Pending
-    }
-}
-
-/// HTTP/3 client builder
-///
-/// Set the configuration for a new client.
-///
-/// # Examples
-/// ```rust
-/// # use h3::quic;
-/// # async fn doc<C, O, B>(quic: C)
-/// # where
-/// #   C: quic::Connection<B, OpenStreams = O>,
-/// #   O: quic::OpenStreams<B>,
-/// #   B: bytes::Buf,
-/// # {
-/// let h3_conn = h3::client::client::builder()
-///     .max_field_section_size(8192)
-///     .build(quic)
-///     .await
-///     .expect("Failed to build connection");
-/// # }
-/// ```
-pub struct Builder {
-    config: Config,
-}
-
-impl Builder {
-    pub(super) fn new() -> Self {
-        Builder {
-            config: Default::default(),
-        }
-    }
-
-    #[cfg(test)]
-    pub fn send_settings(&mut self, value: bool) -> &mut Self {
-        self.config.send_settings = value;
-        self
-    }
-
-    /// Set the maximum header size this client is willing to accept
-    ///
-    /// See [header size constraints] section of the specification for details.
-    ///
-    /// [header size constraints]: https://www.rfc-editor.org/rfc/rfc9114.html#name-header-size-constraints
-    pub fn max_field_section_size(&mut self, value: u64) -> &mut Self {
-        self.config.settings.max_field_section_size = value;
-        self
-    }
-
-    /// Just like in HTTP/2, HTTP/3 also uses the concept of "grease"
-    /// to prevent potential interoperability issues in the future.
-    /// In HTTP/3, the concept of grease is used to ensure that the protocol can evolve
-    /// and accommodate future changes without breaking existing implementations.
-    pub fn send_grease(&mut self, enabled: bool) -> &mut Self {
-        self.config.send_grease = enabled;
-        self
-    }
-
-    /// Create a new HTTP/3 client from a `quic` connection
-    pub async fn build<C, O, B>(
-        &mut self,
-        quic: C,
-    ) -> Result<(Connection<C, B>, SendRequest<O, B>), Error>
-    where
-        C: quic::Connection<B, OpenStreams = O>,
-        O: quic::OpenStreams<B>,
-        B: Buf,
-    {
-        let open = quic.opener();
-        let conn_state = SharedStateRef::default();
-
-        let conn_waker = Some(future::poll_fn(|cx| Poll::Ready(cx.waker().clone())).await);
-
-        Ok((
-            Connection {
-                inner: ConnectionInner::new(quic, conn_state.clone(), self.config).await?,
-                sent_closing: None,
-                recv_closing: None,
-            },
-            SendRequest {
-                open,
-                conn_state,
-                conn_waker,
-                max_field_section_size: self.config.settings.max_field_section_size,
-                sender_count: Arc::new(AtomicUsize::new(1)),
-                send_grease_frame: self.config.send_grease,
-                _buf: PhantomData,
-            },
-        ))
-    }
-}
-
-/// Manage request bodies transfer, response and trailers.
-///
-/// Once a request has been sent via [`send_request()`], a response can be awaited by calling
-/// [`recv_response()`]. A body for this request can be sent with [`send_data()`], then the request
-/// shall be completed by either sending trailers with [`send_trailers()`], or [`finish()`].
-///
-/// After receiving the response's headers, it's body can be read by [`recv_data()`] until it returns
-/// `None`. Then the trailers will eventually be available via [`recv_trailers()`].
-///
-/// TODO: If data is polled before the response has been received, an error will be thrown.
-///
-/// TODO: If trailers are polled but the body hasn't been fully received, an UNEXPECT_FRAME error will be
-/// thrown
-///
-/// Whenever the client wants to cancel this request, it can call [`stop_sending()`], which will
-/// put an end to any transfer concerning it.
-///
-/// # Examples
-///
-/// ```rust
-/// # use h3::{quic, client::client::*};
-/// # use http::{Request, Response};
-/// # use bytes::Buf;
-/// # use tokio::io::AsyncWriteExt;
-/// # async fn doc<T,B>(mut req_stream: RequestStream<T, B>) -> Result<(), Box<dyn std::error::Error>>
-/// # where
-/// #     T: quic::RecvStream,
-/// # {
-/// // Prepare the HTTP request to send to the server
-/// let request = Request::get("https://www.example.com/").body(())?;
-///
-/// // Receive the response
-/// let response = req_stream.recv_response().await?;
-/// // Receive the body
-/// while let Some(mut chunk) = req_stream.recv_data().await? {
-///     let mut out = tokio::io::stdout();
-///     out.write_all_buf(&mut chunk).await?;
-///     out.flush().await?;
-/// }
-/// # Ok(())
-/// # }
-/// # pub fn main() {}
-/// ```
-///
-/// [`send_request()`]: struct.SendRequest.html#method.send_request
-/// [`recv_response()`]: #method.recv_response
-/// [`recv_data()`]: #method.recv_data
-/// [`send_data()`]: #method.send_data
-/// [`send_trailers()`]: #method.send_trailers
-/// [`recv_trailers()`]: #method.recv_trailers
-/// [`finish()`]: #method.finish
-/// [`stop_sending()`]: #method.stop_sending
-pub struct RequestStream<S, B> {
-    inner: connection::RequestStream<S, B>,
-}
-
-impl<S, B> ConnectionState for RequestStream<S, B> {
-    fn shared_state(&self) -> &SharedStateRef {
-        &self.inner.conn_state
-    }
-}
-
-impl<S, B> RequestStream<S, B>
-where
-    S: quic::RecvStream,
-{
-    /// Receive the HTTP/3 response
-    ///
-    /// This should be called before trying to receive any data with [`recv_data()`].
-    ///
-    /// [`recv_data()`]: #method.recv_data
-    pub async fn recv_response(&mut self) -> Result<Response<()>, Error> {
-        let mut frame = future::poll_fn(|cx| self.inner.stream.poll_next(cx))
-            .await
-            .map_err(|e| self.maybe_conn_err(e))?
-            .ok_or_else(|| {
-                Code::H3_GENERAL_PROTOCOL_ERROR.with_reason(
-                    "Did not receive response headers",
-                    ErrorLevel::ConnectionError,
-                )
-            })?;
-
-        //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.5
-        //= type=TODO
-        //# A client MUST treat
-        //# receipt of a PUSH_PROMISE frame that contains a larger push ID than
-        //# the client has advertised as a connection error of H3_ID_ERROR.
-
-        //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.5
-        //= type=TODO
-        //# If a client
-        //# receives a push ID that has already been promised and detects a
-        //# mismatch, it MUST respond with a connection error of type
-        //# H3_GENERAL_PROTOCOL_ERROR.
-
-        let decoded = if let Frame::Headers(ref mut encoded) = frame {
-            match qpack::decode_stateless(encoded, self.inner.max_field_section_size) {
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-4.2.2
-                //# An HTTP/3 implementation MAY impose a limit on the maximum size of
-                //# the message header it will accept on an individual HTTP message.
-                Err(qpack::DecoderError::HeaderTooLong(cancel_size)) => {
-                    self.inner.stop_sending(Code::H3_REQUEST_CANCELLED);
-                    return Err(Error::header_too_big(
-                        cancel_size,
-                        self.inner.max_field_section_size,
-                    ));
-                }
-                Ok(decoded) => decoded,
-                Err(e) => return Err(e.into()),
-            }
-        } else {
-            return Err(Code::H3_FRAME_UNEXPECTED.with_reason(
-                "First response frame is not headers",
-                ErrorLevel::ConnectionError,
-            ));
-        };
-
-        let qpack::Decoded { fields, .. } = decoded;
-
-        let (status, headers) = Header::try_from(fields)?.into_response_parts()?;
-        let mut resp = Response::new(());
-        *resp.status_mut() = status;
-        *resp.headers_mut() = headers;
-        *resp.version_mut() = http::Version::HTTP_3;
-
-        Ok(resp)
-    }
-
-    /// Receive some of the request body.
-    // TODO what if called before recv_response ?
-    pub async fn recv_data(&mut self) -> Result<Option<impl Buf>, Error> {
-        self.inner.recv_data().await
-    }
-
-    /// Receive an optional set of trailers for the response.
-    pub async fn recv_trailers(&mut self) -> Result<Option<HeaderMap>, Error> {
-        let res = self.inner.recv_trailers().await;
-        if let Err(ref e) = res {
-            if e.is_header_too_big() {
-                self.inner.stream.stop_sending(Code::H3_REQUEST_CANCELLED);
-            }
-        }
-        res
-    }
-
-    /// Tell the peer to stop sending into the underlying QUIC stream
-    pub fn stop_sending(&mut self, error_code: crate::error::Code) {
-        // TODO take by value to prevent any further call as this request is cancelled
-        // rename `cancel()` ?
-        self.inner.stream.stop_sending(error_code)
-    }
-}
-
-impl<S, B> RequestStream<S, B>
-where
-    S: quic::SendStream<B>,
-    B: Buf,
-{
-    /// Send some data on the request body.
-    pub async fn send_data(&mut self, buf: B) -> Result<(), Error> {
-        self.inner.send_data(buf).await
-    }
-
-    /// Send a set of trailers to end the request.
-    ///
-    /// Either [`RequestStream::finish`] or
-    /// [`RequestStream::send_trailers`] must be called to finalize a
-    /// request.
-    pub async fn send_trailers(&mut self, trailers: HeaderMap) -> Result<(), Error> {
-        self.inner.send_trailers(trailers).await
-    }
-
-    /// End the request without trailers.
-    ///
-    /// Either [`RequestStream::finish`] or
-    /// [`RequestStream::send_trailers`] must be called to finalize a
-    /// request.
-    pub async fn finish(&mut self) -> Result<(), Error> {
-        self.inner.finish().await
-    }
-
-    //= https://www.rfc-editor.org/rfc/rfc9114#section-4.1.1
-    //= type=TODO
-    //# Implementations SHOULD cancel requests by abruptly terminating any
-    //# directions of a stream that are still open.  To do so, an
-    //# implementation resets the sending parts of streams and aborts reading
-    //# on the receiving parts of streams; see Section 2.4 of
-    //# [QUIC-TRANSPORT].
-}
-
-impl<S, B> RequestStream<S, B>
-where
-    S: quic::BidiStream<B>,
-    B: Buf,
-{
-    /// Split this stream into two halves that can be driven independently.
-    pub fn split(
-        self,
-    ) -> (
-        RequestStream<S::SendStream, B>,
-        RequestStream<S::RecvStream, B>,
-    ) {
-        let (send, recv) = self.inner.split();
-        (RequestStream { inner: send }, RequestStream { inner: recv })
     }
 }
