+++ h3/src/client/mod.rs
@@ -1,2 +1,6 @@
-/// http3 client implementation
-pub mod client;
+//! HTTP/3 client
+
+pub mod connection;
+pub mod stream;
+
+pub mod builder;
