+++ h3/src/client/stream.rs
@@ -1,0 +1,221 @@
+//! Manage request bodies transfer, response and trailers.
+//!
+//! Once a request has been sent via [`send_request()`], a response can be awaited by calling
+//! [`recv_response()`]. A body for this request can be sent with [`send_data()`], then the request
+//! shall be completed by either sending trailers with [`send_trailers()`], or [`finish()`].
+//!
+//! After receiving the response's headers, it's body can be read by [`recv_data()`] until it returns
+//! `None`. Then the trailers will eventually be available via [`recv_trailers()`].
+//!
+//! TODO: If data is polled before the response has been received, an error will be thrown.
+//!
+//! TODO: If trailers are polled but the body hasn't been fully received, an UNEXPECT_FRAME error will be
+//! thrown
+//!
+//! Whenever the client wants to cancel this request, it can call [`stop_sending()`], which will
+//! put an end to any transfer concerning it.
+//!
+//! # Examples
+//!
+//! ```rust
+//! # use h3::{quic, client::connection::*, client::stream::*, client::builder::*};
+//! # use http::{Request, Response};
+//! # use bytes::Buf;
+//! # use tokio::io::AsyncWriteExt;
+//! # async fn doc<T,B>(mut req_stream: RequestStream<T, B>) -> Result<(), Box<dyn std::error::Error>>
+//! # where
+//! #     T: quic::RecvStream,
+//! # {
+//! // Prepare the HTTP request to send to the server
+//! let request = Request::get("https://www.example.com/").body(())?;
+//!
+//! // Receive the response
+//! let response = req_stream.recv_response().await?;
+//! // Receive the body
+//! while let Some(mut chunk) = req_stream.recv_data().await? {
+//!     let mut out = tokio::io::stdout();
+//!     out.write_all_buf(&mut chunk).await?;
+//!     out.flush().await?;
+//! }
+//! # Ok(())
+//! # }
+//! # pub fn main() {}
+//! ```
+//!
+//! [`send_request()`]: struct.SendRequest.html#method.send_request
+//! [`recv_response()`]: #method.recv_response
+//! [`recv_data()`]: #method.recv_data
+//! [`send_data()`]: #method.send_data
+//! [`send_trailers()`]: #method.send_trailers
+//! [`recv_trailers()`]: #method.recv_trailers
+//! [`finish()`]: #method.finish
+//! [`stop_sending()`]: #method.stop_sending
+
+use bytes::Buf;
+use futures_util::future;
+use http::{HeaderMap, Response};
+
+use crate::{
+    connection::{self, ConnectionState, SharedStateRef},
+    error::{Code, Error, ErrorLevel},
+    proto::{frame::Frame, headers::Header},
+    qpack,
+    quic::{self},
+};
+use std::convert::TryFrom;
+
+/// Manage request bodies transfer, response and trailers.
+pub struct RequestStream<S, B> {
+    pub(super) inner: connection::RequestStream<S, B>,
+}
+
+impl<S, B> ConnectionState for RequestStream<S, B> {
+    fn shared_state(&self) -> &SharedStateRef {
+        &self.inner.conn_state
+    }
+}
+
+impl<S, B> RequestStream<S, B>
+where
+    S: quic::RecvStream,
+{
+    /// Receive the HTTP/3 response
+    ///
+    /// This should be called before trying to receive any data with [`recv_data()`].
+    ///
+    /// [`recv_data()`]: #method.recv_data
+    pub async fn recv_response(&mut self) -> Result<Response<()>, Error> {
+        let mut frame = future::poll_fn(|cx| self.inner.stream.poll_next(cx))
+            .await
+            .map_err(|e| self.maybe_conn_err(e))?
+            .ok_or_else(|| {
+                Code::H3_GENERAL_PROTOCOL_ERROR.with_reason(
+                    "Did not receive response headers",
+                    ErrorLevel::ConnectionError,
+                )
+            })?;
+
+        //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.5
+        //= type=TODO
+        //# A client MUST treat
+        //# receipt of a PUSH_PROMISE frame that contains a larger push ID than
+        //# the client has advertised as a connection error of H3_ID_ERROR.
+
+        //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.5
+        //= type=TODO
+        //# If a client
+        //# receives a push ID that has already been promised and detects a
+        //# mismatch, it MUST respond with a connection error of type
+        //# H3_GENERAL_PROTOCOL_ERROR.
+
+        let decoded = if let Frame::Headers(ref mut encoded) = frame {
+            match qpack::decode_stateless(encoded, self.inner.max_field_section_size) {
+                //= https://www.rfc-editor.org/rfc/rfc9114#section-4.2.2
+                //# An HTTP/3 implementation MAY impose a limit on the maximum size of
+                //# the message header it will accept on an individual HTTP message.
+                Err(qpack::DecoderError::HeaderTooLong(cancel_size)) => {
+                    self.inner.stop_sending(Code::H3_REQUEST_CANCELLED);
+                    return Err(Error::header_too_big(
+                        cancel_size,
+                        self.inner.max_field_section_size,
+                    ));
+                }
+                Ok(decoded) => decoded,
+                Err(e) => return Err(e.into()),
+            }
+        } else {
+            return Err(Code::H3_FRAME_UNEXPECTED.with_reason(
+                "First response frame is not headers",
+                ErrorLevel::ConnectionError,
+            ));
+        };
+
+        let qpack::Decoded { fields, .. } = decoded;
+
+        let (status, headers) = Header::try_from(fields)?.into_response_parts()?;
+        let mut resp = Response::new(());
+        *resp.status_mut() = status;
+        *resp.headers_mut() = headers;
+        *resp.version_mut() = http::Version::HTTP_3;
+
+        Ok(resp)
+    }
+
+    /// Receive some of the request body.
+    // TODO what if called before recv_response ?
+    pub async fn recv_data(&mut self) -> Result<Option<impl Buf>, Error> {
+        self.inner.recv_data().await
+    }
+
+    /// Receive an optional set of trailers for the response.
+    pub async fn recv_trailers(&mut self) -> Result<Option<HeaderMap>, Error> {
+        let res = self.inner.recv_trailers().await;
+        if let Err(ref e) = res {
+            if e.is_header_too_big() {
+                self.inner.stream.stop_sending(Code::H3_REQUEST_CANCELLED);
+            }
+        }
+        res
+    }
+
+    /// Tell the peer to stop sending into the underlying QUIC stream
+    pub fn stop_sending(&mut self, error_code: crate::error::Code) {
+        // TODO take by value to prevent any further call as this request is cancelled
+        // rename `cancel()` ?
+        self.inner.stream.stop_sending(error_code)
+    }
+}
+
+impl<S, B> RequestStream<S, B>
+where
+    S: quic::SendStream<B>,
+    B: Buf,
+{
+    /// Send some data on the request body.
+    pub async fn send_data(&mut self, buf: B) -> Result<(), Error> {
+        self.inner.send_data(buf).await
+    }
+
+    /// Send a set of trailers to end the request.
+    ///
+    /// Either [`RequestStream::finish`] or
+    /// [`RequestStream::send_trailers`] must be called to finalize a
+    /// request.
+    pub async fn send_trailers(&mut self, trailers: HeaderMap) -> Result<(), Error> {
+        self.inner.send_trailers(trailers).await
+    }
+
+    /// End the request without trailers.
+    ///
+    /// Either [`RequestStream::finish`] or
+    /// [`RequestStream::send_trailers`] must be called to finalize a
+    /// request.
+    pub async fn finish(&mut self) -> Result<(), Error> {
+        self.inner.finish().await
+    }
+
+    //= https://www.rfc-editor.org/rfc/rfc9114#section-4.1.1
+    //= type=TODO
+    //# Implementations SHOULD cancel requests by abruptly terminating any
+    //# directions of a stream that are still open.  To do so, an
+    //# implementation resets the sending parts of streams and aborts reading
+    //# on the receiving parts of streams; see Section 2.4 of
+    //# [QUIC-TRANSPORT].
+}
+
+impl<S, B> RequestStream<S, B>
+where
+    S: quic::BidiStream<B>,
+    B: Buf,
+{
+    /// Split this stream into two halves that can be driven independently.
+    pub fn split(
+        self,
+    ) -> (
+        RequestStream<S::SendStream, B>,
+        RequestStream<S::RecvStream, B>,
+    ) {
+        let (send, recv) = self.inner.split();
+        (RequestStream { inner: send }, RequestStream { inner: recv })
+    }
+}
