+++ h3/src/lib.rs
@@ -2,8 +2,8 @@
 #![deny(missing_docs, clippy::self_named_module_files)]
 #![allow(clippy::derive_partial_eq_without_eq)]
 
-/// HTTP/3 client
 pub mod client;
+
 mod config;
 pub mod error;
 pub mod ext;
